+++ tools/sponsorsToContentful.py
@@ -18,10 +18,46 @@ tierNameToSlug = {
 	'Gold': 'gold',
 	'Silver': 'silver',
 	'Bronze': 'bronze',
+	'Copper': 'copper',
 	'Automobile': 'automobile',
 	'Custom': 'custom',
-	'Friends & Family': 'friends-family'
+	'Friends & Family': 'friends-family',
+	'Business Card': 'business-card',
+	'Inkind Donation': 'inkind'
 }
+
+allTierSlugs = list(tierNameToSlug.values())
+
+# update enum
+if not contentful.noPublish:
+	ct = contentful.session.get(f'https://api.contentful.com/spaces/{contentful.spaceId}/environments/{contentful.environmentId}/content_types/{contentTypeId}').json()
+
+	for field in ct['fields']:
+		if field['id'] == 'tier':
+			flag = False
+			for valid in field['validations']:
+				if 'in' in valid:
+					valid['in'] = allTierSlugs
+					flag = True
+			if not flag:
+				field['validations'].append({
+					'in': allTierSlugs
+				})
+			break
+
+	version = str(ct['sys']['version'])
+	del ct['sys']
+
+	req = contentful.session.put(f'https://api.contentful.com/spaces/{contentful.spaceId}/environments/{contentful.environmentId}/content_types/{contentTypeId}', json=ct, headers={
+		'X-Contentful-Version': version,
+	})
+
+	version = str(req.json()['sys']['version'])
+
+	# activate version
+	req = contentful.session.put(f'https://api.contentful.com/spaces/{contentful.spaceId}/environments/{contentful.environmentId}/content_types/{contentTypeId}/published', headers={
+		'X-Contentful-Version': version,
+	})
 
 newEntries = {}
 
@@ -29,6 +65,8 @@ with open('sponsors.csv') as csvfile:
 	reader = csv.DictReader(csvfile)
 
 	for sponsor in reader:
+		if not sponsor.get('heading'): continue
+
 		entry = {
 			"metadata": {
 				"tags": [
@@ -48,13 +86,19 @@ with open('sponsors.csv') as csvfile:
 			}
 		}
 
-		for key in ['heading', 'link']:
-			if sponsor.get(key):
-				entry['fields'][key] = {
-					"en-US": sponsor[key].strip()
-				}
-		
 		heading = sponsor['heading'].strip()
+		entry['fields']['heading'] = {
+			"en-US": heading.strip()
+		}
+
+		if sponsor.get('link'):
+			link = sponsor['link'].strip()
+			if link and not link.startswith('http'):
+				link = 'https://' + link
+			entry['fields']['link'] = {
+				"en-US": link
+			}
+
 		newEntries[heading] = {
 			'heading': heading,
 			'entry': entry,
@@ -74,26 +118,27 @@ def doUpdate(heading):
 		# update
 		headers['X-Contentful-Version'] = str(data['version'])
 
-	newData = getattr(contentful.session, 'put' if id else 'post')(f'https://api.contentful.com/spaces/{contentful.spaceId}/environments/{contentful.environmentId}/entries{f"/{id}" if id else ""}', json=data['entry'], headers=headers)
+	if not contentful.noPublish:
+		newData = getattr(contentful.session, 'put' if id else 'post')(f'https://api.contentful.com/spaces/{contentful.spaceId}/environments/{contentful.environmentId}/entries{f"/{id}" if id else ""}', json=data['entry'], headers=headers)
 
-	if newData.status_code != 200 and newData.status_code != 201:
-		print('error')
-		print(data)
-		print(newData.status_code, newData.headers)
-		print(newData.text)
-		print('---')
-		return
+		if newData.status_code != 200 and newData.status_code != 201:
+			print('error')
+			print(data)
+			print(newData.status_code, newData.headers)
+			print(newData.text)
+			print('---')
+			return
 
-	id = newData.json()['sys']['id']
+		id = newData.json()['sys']['id']
 
-	publishPayload.append({
-		'sys': {
-			'id': id,
-			'type': 'Link',
-			'linkType': 'Entry',
-			'version': newData.json()['sys']['version'],
-		}
-	})
+		publishPayload.append({
+			'sys': {
+				'id': id,
+				'type': 'Link',
+				'linkType': 'Entry',
+				'version': newData.json()['sys']['version'],
+			}
+		})
 
 	sleep(0.2)
 
@@ -125,8 +170,9 @@ for entry in prevEntries['items']:
 		headers = {
 			'X-Contentful-Version': str(version),
 		}
-		contentful.session.delete(f'https://api.contentful.com/spaces/{contentful.spaceId}/environments/{contentful.environmentId}/entries/{id}/published', headers=headers)
-		contentful.session.put(f'https://api.contentful.com/spaces/{contentful.spaceId}/environments/{contentful.environmentId}/entries/{id}/archived', headers=headers)
+		if not contentful.noPublish:
+			contentful.session.delete(f'https://api.contentful.com/spaces/{contentful.spaceId}/environments/{contentful.environmentId}/entries/{id}/published', headers=headers)
+			contentful.session.put(f'https://api.contentful.com/spaces/{contentful.spaceId}/environments/{contentful.environmentId}/entries/{id}/archived', headers=headers)
 
 # create the rest that we aren't modifying
 for heading in newEntries:
